backend/requirements.txt
fastapi==0.110.0
gunicorn==21.2.0
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.28
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic[email]==2.6.3
pydantic-settings==2.2.1
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
